'Removing' of gesture feature now added
I was able to implement a feature that will remove the an existing
gesture registered by the app. This is done via 'Remove' button that
will show an input dialog that accepts existing gesture to delete. So
far this thing works. Will test this tomorrow, need to convert this into
excutable for testing. Happy Halloween!