Applying hotfix 1.1.1
This will correctly type the '?' character by using another third party
library PyAutoGUI. The concept is to let the PyAutoGUI function that
will type any meaning that has a question mark and the Keyboard library
will hold that function to trigger it. My idea works 100% without a code
hiccup. Will test this new version for a week before releasing it.