Implemented updateGestures() function in GesturesDatabase class
The app can now update the selected gesture in the table and apply the changes to the database.

As you can see, I'm heavily relying on the list index to access the value I need as a parameter for my functions. This should be in the KeyboardGesture class. I will try to implement this in future coding sessions.