Properly saved the updated gestures to the database
The previous code doesn't get the date_updated value to the database. As you can see in the updateRecord() function in the repository, I'm relying too much indexing the list to get the value I want.

My next challenge is if I can create a collection of keyboardGestures model in keyboardGestureList variable.