Last known size and position of the application is now restored
Silly initial mistakes on my last commit. The window size was not restoring because the _set_properties() is setting the size again back to its default value.

I just moved the _read_settings() after the _set_properties() so that the last known value of the window size will be used. Hello Qt, you got me again! :)