Playing with QSortFilterProxyModel
On initial run, the columns can by sorted by clicking its header. But after adding, updating or deleting a row, the sorting no longer works.

I already played around with this, but I end up crashing the application when I use the sortingProxyModel after a CRUD action.

One thing I haven't tried and still wrapping my head around is the mapToSource and mapFromSource methods of the QSortFilterProxyModel class. This is already advance concept for me so this will take some time to complete.