Delete selected row using mapToSource() method
Once you sort the tableview, it messes the order of your tableview's index. The mapToSource() and mapFromSource() method was created to keep track of the source's model index after sorting. You simply need to pass the tableview's currentIndex() method as argument to these mappings.

So far, that is what I get during this coding session. Pretty interesting concept from Qt.